+++ packages/react-mui-authn/src/features/AuthFlow/utils.tsx
@@ -1,7 +1,11 @@
-import GoogleIcon from "@src/components/Icons/google";
-import GitHubIcon from "@src/components/Icons/github";
-import MicrosoftIcon from "@src/components/Icons/microsoft";
-import FacebookIcon from "@src/components/Icons/facebook";
+import {
+  GoogleIcon,
+  GitHubIcon,
+  MicrosoftIcon,
+  FacebookIcon,
+  GitLabIcon,
+  LinkedInIcon,
+} from "@src/components/Icons";
 
 export const getSocialProviderIcon = (provider: string) => {
   switch (provider) {
@@ -13,6 +17,10 @@ export const getSocialProviderIcon = (provider: string) => {
       return <MicrosoftIcon />;
     case "facebook":
       return <FacebookIcon />;
+    case "gitlab":
+      return <GitLabIcon />;
+    case "linkedin":
+      return <LinkedInIcon />;
     default:
       return <></>;
   }
@@ -28,8 +36,10 @@ export const getSocialProviderLabel = (provider: string): string => {
       return "Microsoft";
     case "facebook":
       return "Facebook";
-    case "webauthn":
-      return "WebAuthn";
+    case "gitlab":
+      return "GitLab";
+    case "linkedin":
+      return "LinkedIn";
     default:
       return provider;
   }
@@ -50,4 +60,38 @@ export const getOtpTitle = (provider: string): string => {
 export const checkForHtml = (value: string | undefined): boolean => {
   const stripped = value?.replace(/(<([^>]+)>)/gi, "");
   return stripped === value;
+};
+
+export const isDark = (color: any): boolean => {
+  // Variables for red, green, blue values
+  var r, g, b, hsp;
+
+  // Check the format of the color, HEX or RGB?
+  if (color.match(/^rgb/)) {
+    // If RGB --> store the red, green, blue values in separate variables
+    color = color.match(
+      /^rgba?\((\d+),\s*(\d+),\s*(\d+)(?:,\s*(\d+(?:\.\d+)?))?\)$/
+    );
+
+    r = color[1];
+    g = color[2];
+    b = color[3];
+  } else {
+    // If hex --> Convert it to RGB: http://gist.github.com/983661
+    color = +("0x" + color.slice(1).replace(color.length < 5 && /./g, "$&$&"));
+
+    r = color >> 16;
+    g = (color >> 8) & 255;
+    b = color & 255;
+  }
+
+  // HSP (Highly Sensitive Poo) equation from http://alienryderflex.com/hsp.html
+  hsp = Math.sqrt(0.299 * (r * r) + 0.587 * (g * g) + 0.114 * (b * b));
+
+  // Using the HSP value, determine whether the color is light or dark
+  if (hsp > 127.5) {
+    return false;
+  } else {
+    return true;
+  }
 };