+++ packages/react-mui-store-file-viewer/src/stories/StoreDownloadFileViewer.stories.tsx
@@ -41,11 +41,13 @@ StoreFileViewerDemo.args = {
     list: async (body) => {
       return axios
         .post(
-          `https://store.${process.env.STORYBOOK_SERVICE_DOMAIN}/v1beta/list`,
+          `https://share.${
+            import.meta.env.STORYBOOK_SERVICE_DOMAIN
+          }/v1beta/list`,
           { ...body },
           {
             headers: {
-              Authorization: `Bearer ${process.env.STORYBOOK_PANGEA_TOKEN}`,
+              Authorization: `Bearer ${import.meta.env.STORYBOOK_PANGEA_TOKEN}`,
             },
           }
         )
@@ -60,11 +62,13 @@ StoreFileViewerDemo.args = {
     get: async (body) => {
       return axios
         .post(
-          `https://store.${process.env.STORYBOOK_SERVICE_DOMAIN}/v1beta/get`,
+          `https://share.${
+            import.meta.env.STORYBOOK_SERVICE_DOMAIN
+          }/v1beta/get`,
           { ...body },
           {
             headers: {
-              Authorization: `Bearer ${process.env.STORYBOOK_PANGEA_TOKEN}`,
+              Authorization: `Bearer ${import.meta.env.STORYBOOK_PANGEA_TOKEN}`,
             },
           }
         )
@@ -79,11 +83,13 @@ StoreFileViewerDemo.args = {
     getArchive: async (body) => {
       return axios
         .post(
-          `https://store.${process.env.STORYBOOK_SERVICE_DOMAIN}/v1beta/get_archive`,
+          `https://share.${
+            import.meta.env.STORYBOOK_SERVICE_DOMAIN
+          }/v1beta/get_archive`,
           { ...body },
           {
             headers: {
-              Authorization: `Bearer ${process.env.STORYBOOK_PANGEA_TOKEN}`,
+              Authorization: `Bearer ${import.meta.env.STORYBOOK_PANGEA_TOKEN}`,
             },
           }
         )