+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -10,7 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documentation to service client configuration.
+- Audit /download_results endpoint support
 - Documentation to service client constructors.
+
+### Fixed
+
+- Put to presigned url. It should just put file in raw, not in form format.
 
 ### Removed